requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
google-generativeai==0.6.0 # Added Google Gemini library